pyproject.toml
# HuggingFace Spaces Configuration
# https://huggingface.co/docs/hub/spaces-sdks-docker

[tool.uv]
# UV package manager configuration
required-version = ">=0.5.0"

[project]
name = "retina-ai"
version = "1.0.0"
description = "Retinitis Pigmentosa Detection AI"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.17.0",
    "timm>=0.9.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "starlette>=0.27.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "plotly>=5.17.0",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    "gunicorn>=21.2.0",
    "pytest>=7.4.0",
    "black>=23.10.0",
    "flake8>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.10.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[spaces]
# HuggingFace Spaces specific configuration
title = "Retina AI - RP Detection"
emoji = "👁️"
colorFrom = "blue"
colorTo = "green"
sdk = "docker"
app_port = 7860
